Rename module so that we can add version module without creating cycle imports
